refactored the models to modular parts
