--- api/db_models/restaurant/model.py
@@ -1,6 +1,0 @@
-""" restaurant model """
-
-
-from datetime import datetime
-
-from api.core.base import declarative_base as db
